Add evaluator sub-module stubs and rename type_environment to environment
Rename hir/type_environment.hpp to hir/environment.hpp and fix all
include paths. Add eval/ stubs for expression, statement, let-binding,
and top-item evaluators, plus environment and evaluator context headers.
