+++ bust/eval/expression_evaluator.cpp
@@ -1,0 +1,20 @@
+//**** Copyright © 2023-2026 Sean Carroll. All rights reserved.
+//*
+//*
+//*  Version : $Header:$
+//*
+//*
+//*  Purpose : Implementation of expression evaluator.
+//*
+//*
+//****************************************************************************
+
+#include "eval/expression_evaluator.hpp"
+
+//****************************************************************************
+namespace bust::eval {
+//****************************************************************************
+
+//****************************************************************************
+} // namespace bust::eval
+//****************************************************************************
