+++ bust/eval/let_binding_evaluator.cpp
@@ -1,0 +1,20 @@
+//**** Copyright © 2023-2026 Sean Carroll. All rights reserved.
+//*
+//*
+//*  Version : $Header:$
+//*
+//*
+//*  Purpose : Implementation of let-binding evaluator.
+//*
+//*
+//****************************************************************************
+
+#include "eval/let_binding_evaluator.hpp"
+
+//****************************************************************************
+namespace bust::eval {
+//****************************************************************************
+
+//****************************************************************************
+} // namespace bust::eval
+//****************************************************************************
