+++ bust/eval/statement_evaluator.cpp
@@ -1,0 +1,20 @@
+//**** Copyright © 2023-2026 Sean Carroll. All rights reserved.
+//*
+//*
+//*  Version : $Header:$
+//*
+//*
+//*  Purpose : Implementation of statement evaluator.
+//*
+//*
+//****************************************************************************
+
+#include "eval/statement_evaluator.hpp"
+
+//****************************************************************************
+namespace bust::eval {
+//****************************************************************************
+
+//****************************************************************************
+} // namespace bust::eval
+//****************************************************************************
