+++ bust/eval/top_item_evaluator.cpp
@@ -1,0 +1,20 @@
+//**** Copyright © 2023-2026 Sean Carroll. All rights reserved.
+//*
+//*
+//*  Version : $Header:$
+//*
+//*
+//*  Purpose : Implementation of top-level item evaluator.
+//*
+//*
+//****************************************************************************
+
+#include "eval/top_item_evaluator.hpp"
+
+//****************************************************************************
+namespace bust::eval {
+//****************************************************************************
+
+//****************************************************************************
+} // namespace bust::eval
+//****************************************************************************
