+++ bust/src/type_checker.cpp
@@ -13,10 +13,10 @@
 #include "ast/types.hpp"
 #include "exceptions.hpp"
 #include "hir/checker_context.hpp"
+#include "hir/environment.hpp"
 #include "hir/nodes.hpp"
 #include "hir/top_item_checker.hpp"
 #include "hir/type_converter.hpp"
-#include "hir/type_environment.hpp"
 #include "hir/type_unifier.hpp"
 #include "hir/type_visitors.hpp"
 #include "hir/types.hpp"
